pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mlops_group_11"
version = "0.0.1"
description = "Description"
authors = [
  { name = "MLOps Group 11", email = "your@email.com" },
]
keywords = ["machine learning", "MLOps"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3",
]
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "torch==2.2.0",
    "torchvision",
    "typer==0.21.1",
    "fastapi==0.128.0",
    "uvicorn==0.34.2",
    "timm>=1.0.24",
    "matplotlib>=3.10.8",
    "hydra-core>=1.3.2",
    "numpy<2.0",
    "scikit-learn>=1.5.0",
    "pillow>=10.0.0",
    "pandas>=2.0.0",
    "kaggle>=1.6.0",
    "dotenv>=0.9.9",
    "wandb>=0.24.0",
    "onnx>=1.19.0",
    "onnxruntime>=1.23.2",
    "onnxscript>=0.5.0",
    "bentoml>=1.4.33",
]

[dependency-groups]
dev = [
    "invoke==2.2.1",
    "pre-commit==4.1.0",
    "pytest==8.3.4",
    "coverage==7.13.1",
    "ruff==0.14.11",
    "mkdocs==1.6.1",
    "mkdocs-material==9.7.1",
    "mkdocstrings-python==2.0.1",
    "mypy>=1.19.1",
    "snakeviz>=2.2.2",
    "torch-tb-profiler>=0.4.3",
]

[tool.uv.sources]
torch = [{ index = "pytorch-cpu" }]
torchvision = [{ index = "pytorch-cpu" }]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.ruff]
line-length = 120
lint.select = ["I"]

[tool.coverage.run]
omit = ["tests/*"]
